se hicieron cambios para utilizar decimal en lugar de double , esperemos que funcione

+++ Application/Interfaces/IRecibo/IReciboService.cs
@@ -9,7 +9,7 @@ namespace Application.Interfaces.IRecibo
 {
     public interface IReciboService
     {
-        ReciboResponse CambiarPrecio(Guid idRecibo, double precioTotal);
+        ReciboResponse CambiarPrecio(Guid idRecibo, decimal precioTotal);
         ReciboResponse CrearRecibo();
         ReciboResponse GetReciboById(Guid id);
         List<ReciboResponse> GetRecibosPersonal(Guid idPersonal);
